+++ packages/python/plotly/plotly/tests/test_optional/test_px/test_colors.py
