+++ packages/python/plotly/test_requirements/requirements_36_core.txt
@@ -1,10 +1,4 @@
 requests==2.12.4
 six==1.10.0
 retrying==1.3.3
-pandas==0.24.2
-numpy==1.19.5
-xarray==0.10.9
-statsmodels==0.10.2
-pillow==5.2.0
 pytest==3.5.1
-pytz==2016.10
